router master performance issue

+++ core/egov-workflow/src/main/java/org/egov/workflow/domain/model/PersistRouter.java
@@ -1,14 +1,8 @@
 package org.egov.workflow.domain.model;
 
-import javax.validation.constraints.NotNull;
+import lombok.*;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
+import javax.validation.constraints.NotNull;
 
 @AllArgsConstructor
 @EqualsAndHashCode
@@ -18,23 +12,24 @@ import lombok.ToString;
 @ToString
 @Builder
 public class PersistRouter {
-	public static final String SEQ_ROUTER = "SEQ_EGPGR_ROUTER";
-	@NotNull
-	private Long id;
-	
-	private Long service;
-	
-	private Integer boundary;
-	
-	private Boolean active; 
-	
-	@NotNull
-	private Integer position;
-	
-	@NotNull
-	private String tenantId;
-	
-	private AuditDetails auditDetails;
-	
+    public static final String SEQ_ROUTER = "SEQ_EGPGR_ROUTER";
+
+    @NotNull
+    private Long id;
+
+    private Long service;
+
+    private Long boundary;
+
+    private Boolean active;
+
+    @NotNull
+    private Integer position;
+
+    @NotNull
+    private String tenantId;
+
+    private AuditDetails auditDetails;
+
 }
